docs: importable sphinx autodoc aliases
Sphinx allows documentation authors to refer to classes, functions, etc.
via a syntax `:py:class:`aiida.orm.Dict`.

Since 8a5d8017e8cd09d7e477aa41c8a07bb24a35e58, using the top-level
`:py:class:`aiida.orm.Dict` was broken and only
via a syntax `:py:class:`aiida.orm.nodes.data.dict.Dict` continued to
work, breaking the documentation of many AiiDA plugins.

8a5d8017e8cd09d7e477aa41c8a07bb24a35e58 included a workaround (via
sphinx aliases), that was, however, only applied to the aiida-core
documentation. Here, we use the AiiDA sphinxextension to make these
aliases available to plugins as well.
